In most of our benchmarks, HTTP validation is not enabled and it doesn't make easy to spot which cost it has.
This PR is addressing this by adding special flags to both test it with/without ASCII values and it improves it by removing branches out of its validation path.