@franz1981 2023-01-17
Preliminary results (not yet inspected Assembly):

master:
```
Benchmark                       (asciiNames)  (validate)   Mode  Cnt      Score     Error   Units
HeadersSetBenchmark.vertxSmall          true        true  thrpt    5   1679.651 ±  19.126  ops/ms
HeadersSetBenchmark.vertxSmall         false        true  thrpt    5   1863.817 ±  21.412  ops/ms
```

This PR at 6cf8aabae7c35ec5b9de1b89c785d179acc6fbaf
```
Benchmark                       (asciiNames)  (validate)   Mode  Cnt      Score      Error   Units
HeadersSetBenchmark.vertxSmall          true        true  thrpt    5   3588.305 ±   52.113  ops/ms
HeadersSetBenchmark.vertxSmall         false        true  thrpt    5   2578.615 ±   33.672  ops/ms
```

The relevant parts are `validate = true`  and `vertx`, while these are the whole comparison 
with Netty at 6cf8aabae7c35ec5b9de1b89c785d179acc6fbaf:
```
Benchmark                       (asciiNames)  (validate)   Mode  Cnt      Score      Error   Units
HeadersSetBenchmark.nettySmall          true        true  thrpt    5   3214.537 ±   67.707  ops/ms
HeadersSetBenchmark.nettySmall          true       false  thrpt    5  17832.882 ±  951.593  ops/ms
HeadersSetBenchmark.nettySmall         false        true  thrpt    5   2555.200 ±   82.710  ops/ms
HeadersSetBenchmark.nettySmall         false       false  thrpt    5   8434.778 ±  118.626  ops/ms
HeadersSetBenchmark.vertxSmall          true        true  thrpt    5   3588.305 ±   52.113  ops/ms
HeadersSetBenchmark.vertxSmall          true       false  thrpt    5  18656.050 ±  140.765  ops/ms
HeadersSetBenchmark.vertxSmall         false        true  thrpt    5   2578.615 ±   33.672  ops/ms
HeadersSetBenchmark.vertxSmall         false       false  thrpt    5   8850.321 ± 1057.573  ops/ms
```